private folders created and moved components
