+++ src/app/(authentication)/layout.tsx
