+++ src/app/(authentication)/tickets/[ticketId]/not-found.tsx
