+++ src/app/(authentication)/tickets/[ticketId]/page.tsx
