+++ src/app/(authentication)/tickets/error.tsx
