+++ src/app/_navigation/AccountDropdown.tsx
@@ -1,7 +1,8 @@
 import { signOut } from "@/features/auth/actions/sign-out";
-import SubmitButton from "./form/SubmitButton";
 import { User } from "@prisma/client";
 import { LucideLogOut, LucideUser } from "lucide-react";
+import Link from "next/link";
+import { accountPasswordPath, accountProfilePath } from "@/paths";
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -9,10 +10,8 @@ import {
   DropdownMenuLabel,
   DropdownMenuSeparator,
   DropdownMenuTrigger,
-} from "./ui/dropdown-menu";
-import { Avatar, AvatarFallback } from "./ui/avatar";
-import Link from "next/link";
-import { accountPasswordPath, accountProfilePath } from "@/paths";
+} from "@/components/ui/dropdown-menu";
+import { Avatar, AvatarFallback } from "@/components/ui/avatar";
 
 type AccountDropdownProps = {
   user: User;
